app/requirements.txt
# FastAPI и сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0

# База данных
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Аутентификация и безопасность
PyJWT==2.8.0
bcrypt==4.1.2
python-bcrypt==4.1.2

# Валидация и настройки
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-dotenv==1.0.0
